ENH: creating a compile time option for the KNN library
It was hard-coded to NOT allow self matches during KNN search, required by the KNNGraphAlphaMutualInformation metric. This may not be desirable.

Now it is configurable through the CMake option ELASTIX_KNN_ALLOW_SELF_MATCH, which defaults to false for backward compatibility.

It is still a compile time option and not a run time option.


git-svn-id: https://svn.bigr.nl/elastix/trunkpublic@4024 4d348492-3962-11de-aa85-65a2020054a3